% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sorensenDice.R
\name{bigrams}
\alias{bigrams}
\title{Calculate the bigrams}
\usage{
bigrams(string)
}
\arguments{
\item{string}{a string for calculating the bigrams}
}
\value{
the bigrams of the string
}
\description{
Calculate the bigrams of a string
}
\examples{
bigrams('saturday')
}
\author{
Daniel Rodriguez Perez
}

