% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FinancialSanctions.R
\docType{package}
\name{FinancialSanctions-package}
\alias{FinancialSanctions-package}
\title{FinancialSanctions}
\description{
FinancialSanctions
}
\details{
FinancialSanctions

\tabular{ll}{
  Package: \tab FinancialSanctions \cr
  Version: \tab 0.0.1 \cr
  License: \tab GPL-3 \cr
}
}
\author{
Daniel Rodriguez Perez \email{daniel.rodriguez.perez@gmail.com}
}
\references{
Mi referencias
}

