% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getSanctionsHMTreasury}
\alias{getSanctionsHMTreasury}
\title{Get the consolitate list of santions from HM Treasury}
\usage{
getSanctionsHMTreasury(file = NULL)
}
\arguments{
\item{file}{the URL to the file}
}
\value{
a santions objec
}
\description{
Get the consolidate list of santions from the HM Treasury
}
\author{
Daniel Rodriguez Perez
}

