% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jaroWinkler.R
\name{jaroWinkler}
\alias{jaroWinkler}
\title{Calculate Jaro-Winkler distance}
\usage{
jaroWinkler(str_1, str_2, weight_threshold = 0.7, num_chars = 4)
}
\arguments{
\item{str_1}{first string for calculating the distance}

\item{str_2}{second string for calculating the distance}

\item{weight_threshold}{percent to apply the Winkler modification}

\item{num_chars}{size of the prefix to be concidered by the Winkler modification}
}
\value{
the Jaro-Winkler distance between the specified strings
}
\description{
Calculate Jaro-Winkler distance for two strings \code{str_1} and
\code{str_2}
}
\examples{
jaroWinkler('saturday', 'sunday')
}
\author{
Daniel Rodriguez Perez
}

