% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/levenshtein.R
\name{levenshtein}
\alias{levenshtein}
\title{Calculate Levenshtein distance}
\usage{
levenshtein(str_1, str_2)
}
\arguments{
\item{str_1}{first string for calculating the distance}

\item{str_2}{second string for calculating the distance}
}
\value{
the Levenshtein distance of two strings
}
\description{
Calculate Levenshtein distance for two strings str_1 and str_2
}
\details{
The Levenshtein distance is the minimum number of operations required to
transform one string into another. Operation is defined as either an
insertion, deletion or substitution.
}
\examples{
levenshtein('saturday', 'sunday')
}
\author{
Daniel Rodriguez Perez
}
\references{
"Levenshtein distance." Wikipedia: The Free Encyclopedia. Wikimedia
Foundation, Inc., 27 Sep. 2015.
<http://en.wikipedia.org/wiki/Levenshtein_distance>
}

