% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levenshtein.R
\name{levenshtein_distance}
\alias{levenshtein_distance}
\title{Calculate Levenshtein distance}
\usage{
levenshtein_distance(str_1, str_2)
}
\arguments{
\item{str_1}{first string for calculating the distance}

\item{str_2}{second string for calculating the distance}
}
\value{
the Levenshtein distance between the specified strings
}
\description{
Calculate Levenshtein distance for two strings \code{str_1} and \code{str_2}
}
\details{
The Levenshtein distance is the minimum number of operations required to
transform one string into another. Operation is defined as either an
insertion, deletion or substitution.
}
\examples{
levenshtein_distance('saturday', 'sunday')

}
\author{
Daniel Rodriguez Perez
}
\references{
"Levenshtein distance." Wikipedia: The Free Encyclopedia. Wikimedia
Foundation, Inc., 27 Sep. 2015.
<http://en.wikipedia.org/wiki/Levenshtein_distance>
}

