% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/levenshtein.R
\name{levenshtein_similarity}
\alias{levenshtein_similarity}
\title{Calculate the similarity of Levenshtein distance}
\usage{
levenshtein_similarity(str_1, str_2)
}
\arguments{
\item{str_1}{first string for calculating the distance}

\item{str_2}{second string for calculating the distance}
}
\value{
the similarity based on the Levenshtein distance of two strings
}
\description{
Calculate the similarity of Levenshtein distance for two strings str_1 and
str_2
}
\examples{
levenshtein_similarity('saturday', 'sunday')
}
\author{
Daniel Rodriguez Perez
}

