% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundex.R
\name{soundex}
\alias{soundex}
\title{Calculate the similarity based on the Soundex distance}
\usage{
soundex(str_1, str_2, algorithm = soundex_algorithm("en"))
}
\arguments{
\item{str_1}{first string for calculating the similarity}

\item{str_2}{second string for calculating the similarity}

\item{algorithm}{object with the soundex algorithm}
}
\value{
the similarity based on the Soundex distance between the specified
        strings
}
\description{
Calculate the similarity for two strings \code{str_1} and \code{str_2} based
on the Soundex distance
}
\examples{
soundex('saturday', 'sunday')

}
\author{
Daniel Rodriguez Perez
}

