% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundex.R
\name{soundex_algorithm}
\alias{soundex_algorithm}
\title{Soundex algorithm for the specified language}
\usage{
soundex_algorithm(lang = "en")
}
\arguments{
\item{lang}{the language code}
}
\value{
an object to be used in the Soundex funtions
}
\description{
Returns an object that can be used in the Soundex funtions to execute in
different languages.
}
\details{
This function returns the soundex code algorithms for use in different
languages. The languages currently supported are:
\itemize{
 \item{\code{en} English}
}
}
\examples{
algorithm <- soundex_algorithm(lang = 'en')
soundex_code('Soundex', algorithm = algorithm)

}
\author{
Daniel Rodriguez Perez
}

