% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/soundex.R
\name{soundex_code}
\alias{soundex_code}
\title{Calculate the Soundex code}
\usage{
soundex_code(string)
}
\arguments{
\item{string}{a string or an array of strings for calculating the code}
}
\value{
the Soundex code
}
\description{
Calculate the Soundex code for a string
}
\details{
Soundex is a phonetic algorithm for indexing names by sound. The Soundex
code for a name consists of a letter followed by three numerical digits: the
letter is the first letter of the name, and the digits encode the remaining
consonants. Consonants at a similar place of articulation share the same
digit so.
}
\examples{
soundex_code('Soundex')
}
\author{
Daniel Rodriguez Perez
}

