% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/soundex.R
\name{soundex_difference}
\alias{soundex_difference}
\title{Calculate the Soundex difference}
\usage{
soundex_difference(str_1, str_2)
}
\arguments{
\item{str_1}{first string for calculating the distance}

\item{str_2}{second string for calculating the distance}
}
\value{
an integer that indicates the difference between the Soundex codes
}
\description{
Calculate the Soundex difference for two strings \code{str_1} and \code{str_2}
}
\details{
The returned value is the number of characters in the Soundex values that
are the same. The return value ranges from 0 through 4: 0 indicates weak or
no similarity, and 4 indicates strong similarity or the same values.
}
\examples{
soundex_difference('saturday', 'sunday')
}
\author{
Daniel Rodriguez Perez
}

